+++ include/sqlite_orm/sqlite_orm.h
@@ -1800,6 +1800,8 @@ namespace sqlite_orm {
             unique_t(columns_tuple columns_) : columns(std::move(columns_)) {}
         };
 
+        struct unindexed_t {};
+
         /**
          *  DEFAULT constraint class.
          *  T is a value type.
@@ -2131,6 +2133,7 @@ namespace sqlite_orm {
                                                              check_if<is_foreign_key>,
                                                              check_if_is_type<null_t>,
                                                              check_if_is_type<not_null_t>,
+                                                             check_if_is_type<unindexed_t>,
                                                              check_if_is_template<unique_t>,
                                                              check_if_is_template<default_t>,
                                                              check_if_is_template<check_t>,
@@ -2172,6 +2175,13 @@ namespace sqlite_orm {
 
     inline internal::unique_t<> unique() {
         return {{}};
+    }
+
+    /**
+     *  UNINDEXED constraint builder function. Used in FTS virtual tables.
+     */
+    inline internal::unindexed_t unindexed() {
+        return {};
     }
 
     template<class... Cs>
@@ -3238,7 +3248,7 @@ namespace sqlite_orm {
             bool replace_bindable_with_question = false;
             bool skip_table_name = true;
             bool use_parentheses = true;
-            bool skip_types_and_constraints = false;
+            bool skip_types_and_constraints_except_unindexed = false;
         };
 
         template<class DBOs>
@@ -18215,7 +18225,7 @@ namespace sqlite_orm {
             template<class Ctx>
             auto serialize(const statement_type& statement, const Ctx& context, const std::string& tableName) {
                 std::stringstream ss;
-                ss << "CREATE TABLE " << streaming_identifier(tableName) << " ( "
+                ss << "CREATE TABLE " << streaming_identifier(tableName) << " ("
                    << streaming_expressions_tuple(statement.elements, context) << ")";
                 if(statement_type::is_without_rowid_v) {
                     ss << " WITHOUT ROWID";
@@ -19110,6 +19120,16 @@ namespace sqlite_orm {
         };
 
         template<>
+        struct statement_serializer<unindexed_t, void> {
+            using statement_type = unindexed_t;
+
+            template<class Ctx>
+            std::string operator()(const statement_type& c, const Ctx& context) const {
+                return "UNINDEXED";
+            }
+        };
+
+        template<>
         struct statement_serializer<collate_constraint_t, void> {
             using statement_type = collate_constraint_t;
 
@@ -19204,12 +19224,17 @@ namespace sqlite_orm {
 
                 std::stringstream ss;
                 ss << streaming_identifier(column.name);
-                if(!context.skip_types_and_constraints) {
+                if(!context.skip_types_and_constraints_except_unindexed) {
                     ss << " " << type_printer<field_type_t<column_type>>().print();
                     ss << streaming_column_constraints(
                         call_as_template_base<column_constraints>(polyfill::identity{})(column),
                         column.is_not_null(),
                         context);
+                } else {
+                    using constraints_tuple = typename column_type::constraints_type;
+                    if(tuple_has_type<constraints_tuple, unindexed_t>::value) {
+                        ss << " UNINDEXED";
+                    }
                 }
                 return ss.str();
             }
@@ -19822,7 +19847,7 @@ namespace sqlite_orm {
                 std::stringstream ss;
                 ss << "USING FTS5(";
                 auto subContext = context;
-                subContext.skip_types_and_constraints = true;
+                subContext.skip_types_and_constraints_except_unindexed = true;
                 ss << streaming_expressions_tuple(statement.columns, subContext) << ")";
                 return ss.str();
             }
